Update to run against Puppet 4 vagrant box
The current generation of Puppet Labs vagrant boxes on the former
"Vagrant Cloud" are now only providing Puppet 4. This code will largely
work just fine with Puppet 4 however Vagrant needs to a different method
to run the Puppet code locally. This is due to the fact the Puppet 4 now
uses only directory environments and nothing else.

This commit updates a few items in response to the change to Puppet 4.
First, it locks to a new version of the Vagrant box so hopefully this
will example will run in a good place. Second, it uses the new Puppet 4
related options for the "puppet apply" Vagrant provisioner.